Removed setuptools based installation files
