+++ config/repository_config.json
@@ -13,7 +13,5 @@
    "DEVELOPMENTSTATUS"          : "Development Status :: 4 - Beta",
    "INTENDEDAUDIENCE"           : "Intended Audience :: Developers",
    "TOPIC"                      : "Topic :: Software Development",
-   "INSTALLREQUIRES"            : [],
-   "PACKAGEDATA"                : ["*.pdf"],
    "PACKAGEDOC"                 : "./packagedoc"
 }
