+++ genpackagedoc.py
@@ -1,6 +1,6 @@
 # **************************************************************************************************************
 #
-#  Copyright 2020-2023 Robert Bosch GmbH
+#  Copyright 2020-2026 Robert Bosch GmbH
 #
 #  Licensed under the Apache License, Version 2.0 (the "License");
 #  you may not use this file except in compliance with the License.
@@ -18,11 +18,11 @@
 #
 # genpackagedoc.py
 #
-# XC-CT/ECA3-Queckenstedt
+# XC-HWP/ESW3-Queckenstedt
 #
 # --------------------------------------------------------------------------------------------------------------
 #
-# 30.06.2025
+# 06.03.2026
 #
 # --------------------------------------------------------------------------------------------------------------
 
@@ -64,7 +64,7 @@ except Exception as ex:
     sys.exit(ERROR)
 
 # -- setting up the GenPackageDoc configuration
-oGenPackageDocConfig = None
+oPackageDocConfig = None
 try:
     oPackageDocConfig = CPackageDocConfig(oRepositoryConfig)
 except Exception as ex:
@@ -72,6 +72,9 @@ except Exception as ex:
     printexception(str(ex))
     print()
     sys.exit(ERROR)
+
+DOCBUILDERFULLNAME = oPackageDocConfig.Get("DOCBUILDERFULLNAME")
+print(f"\nThis is {DOCBUILDERFULLNAME}\n")
 
 # -- setting up and calling the doc builder
 try:
