+++ setup/setup.json
@@ -1,0 +1,5 @@
+{
+    "setup_steps" : {
+                        "GENPACKAGEDOC" : "./genpackagedoc.py"
+                    }
+}
